Finish converting badge to TS
