+++ src/components/badge/beta_badge/beta_badge.tsx
@@ -1,0 +1,95 @@
+import React, { FunctionComponent, HTMLAttributes, ReactNode } from 'react';
+import classNames from 'classnames';
+import { CommonProps } from '../../common';
+
+import { EuiToolTip, ToolTipPositions } from '../../tool_tip';
+
+import { EuiIcon, IconType } from '../../icon';
+
+export interface EuiBetaBadgeProps {
+  /**
+   * Supply an icon type if the badge should just be an icon
+   */
+  iconType?: IconType;
+
+  /**
+   * One word label like "Beta" or "Lab"
+   */
+  label: ReactNode;
+
+  /**
+   * Content for the tooltip
+   */
+  tooltipContent?: ReactNode;
+
+  /**
+   * Custom position of the tooltip
+   */
+  tooltipPosition?: ToolTipPositions;
+
+  /**
+   * Optional title will be supplied as tooltip title or title attribute
+   * otherwise the label will be used
+   */
+  title?: string;
+}
+
+type Props = CommonProps & HTMLAttributes<HTMLSpanElement> & EuiBetaBadgeProps;
+
+export const EuiBetaBadge: FunctionComponent<Props> = ({
+  className,
+  label,
+  tooltipContent,
+  tooltipPosition = 'top',
+  title,
+  iconType,
+  ...rest
+}) => {
+  const classes = classNames(
+    'euiBetaBadge',
+    {
+      'euiBetaBadge--iconOnly': iconType,
+    },
+    className
+  );
+
+  let icon;
+  if (iconType) {
+    icon = (
+      <EuiIcon
+        className="euiBetaBadge__icon"
+        type={iconType}
+        size="m"
+        aria-hidden="true"
+      />
+    );
+  }
+
+  if (tooltipContent) {
+    return (
+      <EuiToolTip
+        position={tooltipPosition}
+        content={tooltipContent}
+        title={title || label}>
+        <span className={classes} {...rest}>
+          {icon || label}
+        </span>
+      </EuiToolTip>
+    );
+  } else {
+    const spanTitle = title || label;
+    if (spanTitle && typeof spanTitle !== 'string') {
+      throw new Error(
+        `Only string titles are permitted on badges that do not use tooltips. Found: ${typeof spanTitle}`
+      );
+    }
+    return (
+      <span
+        className={classes}
+        title={spanTitle as string | undefined}
+        {...rest}>
+        {icon || label}
+      </span>
+    );
+  }
+};
