+++ src/components/badge/beta_badge/index.ts
