+++ src/components/badge/index.ts
