+++ src/components/tool_tip/index.ts
@@ -1,3 +1,3 @@
-export { EuiToolTip } from './tool_tip';
+export { EuiToolTip, ToolTipPositions } from './tool_tip';
 
 export { EuiIconTip } from './icon_tip';
